vhdl: refactor and make space for a parser
Move what is currently in `parser` into a new module `syntax/lexer`, and
add a module `syntax/parser` for upcoming work.
