+++ src/vhdl/syntax/lexer/categorizer.rs
