+++ src/vhdl/syntax/parser/mod.rs
@@ -1,0 +1,3 @@
+// Copyright (c) 2017 Fabian Schuiki
+
+//! A VHDL parser.
